pyproject.toml
[project]
name = "wiki-speedrun"
version = "0.1.0"
description = "Wikipedia speedrun benchmark comparing AI pathfinding agents"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Jack Lutz"}
]
keywords = ["wikipedia", "benchmark", "pathfinding", "llm", "embeddings"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core data handling
    "numpy>=1.24.0",
    "msgpack>=1.0.0",

    # Graph algorithms
    "networkx>=3.0",

    # Fast similarity search
    "faiss-cpu>=1.7.0",

    # Live embeddings
    "sentence-transformers>=2.2.0",

    # Web requests and scraping
    "requests>=2.28.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",  # Fast HTML parser for BS4

    # Browser automation for visualization
    "playwright>=1.40.0",

    # Web UI
    "streamlit>=1.28.0",
    "plotly>=5.18.0",
    "pyvis>=0.3.0",

    # Utilities
    "tqdm>=4.65.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.urls]
Homepage = "https://github.com/jacklutz/wiki_speedrun"
Repository = "https://github.com/jacklutz/wiki_speedrun"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "ui*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # Pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "W",      # pycodestyle warnings
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["src", "ui"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
